+++ .speakeasy/logs/changes/changes.md
@@ -1,4 +1,4 @@
-## Python SDK Changes:
+## Typescript SDK Changes:
 * `outpost.publish()`: **Added**
 * `outpost.retry()`: **Added**
 * `outpost.publish.event()`: **Removed** (Breaking ⚠️)